feat: refactor rename UpdateProjectForm folder
