+++ src/pages/Projects/UpdateProjectModal/index.tsx
@@ -1,6 +1,6 @@
 import {FC} from "react";
 import {ModalTemplate} from "@view/ModalTemplate";
-import {UpdateProjectForm} from "../UpdateProjectFrom";
+import {UpdateProjectForm} from "@/pages/Projects/UpdateProjectForm";
 import {ProjectFormFields} from "@/models/FormFieldsTypes";
 import {ModalContextTemplateProvider} from "@/components/view/ModalTemplate/ModalTemplateContext";
